+++ packages/koop-core/src/models/dataset.js
@@ -2,6 +2,8 @@ function Dataset (koop) {
   this.koop = koop
 }
 
+// Data Operations
+
 Dataset.prototype.getFromCache = function (key, options, callback) {
   this.koop.cache.retrieve(key, options, callback)
 }
@@ -21,6 +23,8 @@ Dataset.prototype.upsertIntoCache = function (key, geojson, options, callback) {
 Dataset.prototype.deleteFromCache = function (key, callback) {
   this.koop.cache.delete(key, callback)
 }
+
+// Metadata Operations
 
 Dataset.prototype.getFromCatalog = function (key, options, callback) {
   this.koop.cache.catalog.retrieve(key, options, callback)