+++ packages/koop-core/CHANGELOG.md
@@ -2,6 +2,13 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+### Added
+* Providers can leverage the cache implicitly
+
+### Changed
+* Base Model inherits from Provider Model
+
 ## [3.3.0] - 2017-03-02
 ### Added
 * Dataset model `upsert`